+++ src/apps/LavaMail/controllers/CtrlThreadList.js
@@ -132,7 +132,10 @@ module.exports = ($rootScope, $scope, $state, $timeout, $interval, $translate,
 	};
 
 	$scope.replyThread = (event, tid) => {
-		$scope.showPopup('compose', {replyThreadId: tid});
+		let thread = $scope.threads[tid];
+		console.log(thread);
+		if (thread)
+			$scope.showPopup('compose', {replyThreadId: tid, replyEmailId: thread.emails[thread.emails.length - 1]});
 	};
 
 	$scope.deleteThread = (tid) => {